+++ examples/python/examples.py
@@ -108,9 +108,9 @@ def commands():
             'help': "Deletes a card from a controller",
             'fn': delete_card,
         },
-        'delete-cards': {
+        'delete-all-cards': {
             'help': "Deletes all cards from a controller",
-            'fn': delete_cards,
+            'fn': delete_all_cards,
         },
         'get-event-index': {
             'help': "Retrieves the current event index from a controller",
@@ -168,6 +168,14 @@ def commands():
             'help': "Sets the supervisor passcodes for keypad-only access to a door",
             'fn': set_door_passcodes,
         },
+        'get-antipassback': {
+            'help': "Retrieves the controller anti-passback mode",
+            'fn': get_antipassback,
+        },
+        'set-antipassback': {
+            'help': "Sets the controller anti-passback mode",
+            'fn': set_antipassback,
+        },
         'restore-default-parameters': {
             'help': "Resets a controller to the manufacturer default configuration",
             'fn': restore_default_parameters,
@@ -179,32 +187,217 @@ def commands():
     }
 
 
+# def parse_args():
+#     parser = argparse.ArgumentParser(description='uhppoted-dll examples')
+#
+#     parser.add_argument('command', type=str, help='command')
+#     parser.add_argument('--controller', type=int, default=DEVICE_ID, help='controller serial number')
+#     parser.add_argument('--interlock', type=int, default=1, help='controller door interlock')
+#     parser.add_argument('--ip-address', type=str, default='192.168.1.100', help='controller IP address')
+#     parser.add_argument('--subnet-mask', type=str, default='255.255.255.0', help='controller subnet mask')
+#     parser.add_argument('--gateway-address', type=str, default='192.168.1.5', help='controller gateway address')
+#     parser.add_argument('--listener-address',
+#                         type=str,
+#                         default='192.168.1.100:60001',
+#                         help='controller event listener address')
+#     parser.add_argument('--listener-interval', type=int, default=0, help='controller event listener auto-send interval')
+#
+#     parser.add_argument('--card', type=int, default=CARD_NUMBER, help='card number')
+#     parser.add_argument('--card-index', type=int, default=CARD_INDEX, help='card index')
+#     parser.add_argument('--door', type=int, default=DOOR, help='controller door ID [1..4]')
+#     parser.add_argument('--event-index', type=int, default=EVENT_INDEX, help='event index')
+#     parser.add_argument('--time-profile', type=int, default=TIME_PROFILE_ID, help='time profile ID')
+#
+#     return parser.parse_args()
+
+
+def parse_args():
+    parser = argparse.ArgumentParser(description='uhppoted-dll examples')
+
+    parser.add_argument('--bind', type=str, default='0.0.0.0', help='UDP IPv4 bind address. Defaults to 0.0.0.0:0')
+
+    parser.add_argument('--broadcast',
+                        type=str,
+                        default='255.255.255.255:60000',
+                        help='UDP IPv4 broadcast address. Defaults to 255.255.255.255:60000')
+
+    parser.add_argument('--listen',
+                        type=str,
+                        default='0.0.0.0:60001',
+                        help='UDP IPv4 event listener bind address. Defaults to 0.0.0.0:60001')
+
+    parser.add_argument('--debug',
+                        action=argparse.BooleanOptionalAction,
+                        default=False,
+                        help='displays sent and received packets')
+
+    parser.add_argument('--destination',
+                        '-d',
+                        type=str,
+                        default=None,
+                        help='(optional) controller IPv4 address:port. Defaults to broadcast address.')
+
+    parser.add_argument('--timeout',
+                        '-t',
+                        type=float,
+                        default=2.5,
+                        help='(optional) operation timeout (in seconds). Defaults to 2.5.')
+
+    parser.add_argument('--udp', action=argparse.BooleanOptionalAction, default=False, help='use UDP protocol')
+    parser.add_argument('--tcp', action=argparse.BooleanOptionalAction, default=False, help='use TCP protocol')
+
+    # ... command specific args
+    subparsers = parser.add_subparsers(title='subcommands', dest='command')
+    parsers = {}
+
+    for c in commands().keys():
+        parsers[f'{c}'] = subparsers.add_parser(f'{c}')
+
+    # ... get-controller
+    get_controller = parsers['get-controller']
+    get_controller.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... set-address
+    set_ip = parsers['set-address']
+    set_ip.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-status
+    get_status = parsers['get-status']
+    get_status.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-time
+    get_time = parsers['get-time']
+    get_time.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... set-time
+    set_time = parsers['set-time']
+    set_time.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-listener
+    get_listener = parsers['get-listener']
+    get_listener.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... set-listener
+    set_listener = parsers['set-listener']
+    set_listener.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-door-control
+    get_door_control = parsers['get-door-control']
+    get_door_control.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... set-door-control
+    set_door_control = parsers['set-door-control']
+    set_door_control.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... open-door
+    open_door = parsers['open-door']
+    open_door.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-cards
+    get_cards = parsers['get-cards']
+    get_cards.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-card
+    get_card = parsers['get-card']
+    get_card.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-card-by-index
+    get_card_by_index = parsers['get-card-by-index']
+    get_card_by_index.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... put-card
+    put_card = parsers['put-card']
+    put_card.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... delete-card
+    delete_card = parsers['delete-card']
+    delete_card.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... delete-all-cards
+    delete_all_cards = parsers['delete-all-cards']
+    delete_all_cards.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-event-index
+    get_event_index = parsers['get-event-index']
+    get_event_index.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... set-event-index
+    set_event_index = parsers['set-event-index']
+    set_event_index.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-event
+    get_event = parsers['get-event']
+    get_event.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... record-special-events
+    record_special_events = parsers['record-special-events']
+    record_special_events.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-time-profile
+    get_time_profile = parsers['get-time-profile']
+    get_time_profile.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... set-time-profile
+    set_time_profile = parsers['set-time-profile']
+    set_time_profile.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... clear-time-profiles
+    clear_time_profiles = parsers['clear-time-profiles']
+    clear_time_profiles.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... add-task
+    add_task = parsers['add-task']
+    add_task.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... refresh-tasklist
+    refresh_tasklist = parsers['refresh-tasklist']
+    refresh_tasklist.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... clear-tasklist
+    clear_tasklist = parsers['clear-tasklist']
+    clear_tasklist.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... set-pc-control
+    set_pc_control = parsers['set-pc-control']
+    set_pc_control.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... set-interlock
+    set_interlock = parsers['set-interlock']
+    set_interlock.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... activate-keypads
+    activate_keypads = parsers['activate-keypads']
+    activate_keypads.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... set-door-passcodes
+    set_door_passcodes = parsers['set-door-passcodes']
+    set_door_passcodes.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... get-antipassback
+    get_antipassback = parsers['get-antipassback']
+    get_antipassback.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    # ... set-antipassback
+    set_antipassback = parsers['set-antipassback']
+    set_antipassback.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+    set_antipassback.add_argument('--antipassback',
+                                  type=str,
+                                  default='disabled',
+                                  help='disabled, (1:2);(3:4), (1,3):(2,4), 1:(2,3) or 1:(2,3,4)')
+
+    # ... restore-default-parameters
+    restore_default_parameters = parsers['restore-default-parameters']
+    restore_default_parameters.add_argument('--controller', type=int, help='controller serial number, e.g. 405419896')
+
+    return parser.parse_args()
+
+
 def main():
     if len(sys.argv) < 2:
         usage()
         return -1
 
-    parser = argparse.ArgumentParser(description='uhppoted-dll examples')
-
-    parser.add_argument('command', type=str, help='command')
-    parser.add_argument('--controller', type=int, default=DEVICE_ID, help='controller serial number')
-    parser.add_argument('--interlock', type=int, default=1, help='controller door interlock')
-    parser.add_argument('--ip-address', type=str, default='192.168.1.100', help='controller IP address')
-    parser.add_argument('--subnet-mask', type=str, default='255.255.255.0', help='controller subnet mask')
-    parser.add_argument('--gateway-address', type=str, default='192.168.1.5', help='controller gateway address')
-    parser.add_argument('--listener-address',
-                        type=str,
-                        default='192.168.1.100:60001',
-                        help='controller event listener address')
-    parser.add_argument('--listener-interval', type=int, default=0, help='controller event listener auto-send interval')
-
-    parser.add_argument('--card', type=int, default=CARD_NUMBER, help='card number')
-    parser.add_argument('--card-index', type=int, default=CARD_INDEX, help='card index')
-    parser.add_argument('--door', type=int, default=DOOR, help='controller door ID [1..4]')
-    parser.add_argument('--event-index', type=int, default=EVENT_INDEX, help='event index')
-    parser.add_argument('--time-profile', type=int, default=TIME_PROFILE_ID, help='time profile ID')
-
-    args = parser.parse_args()
+    args = parse_args()
     cmd = args.command
 
     print(args)
@@ -522,7 +715,7 @@ def delete_card(u, args):
     ])
 
 
-def delete_cards(u, args):
+def delete_all_cards(u, args):
     device_id = args.controller
 
     u.delete_cards(device_id)
@@ -757,6 +950,39 @@ def set_door_passcodes(u, args):
         ('passcode 2', passcode2),
         ('passcode 3', passcode3),
         ('passcode 4', passcode4),
+    ])
+
+
+def get_antipassback(u, args):
+    controller = args.controller
+    antipassback = u.get_antipassback(controller)
+
+    display('get-event-index', [
+        ('ID', controller),
+        ('anti-passback', antipassback),
+    ])
+
+
+def set_antipassback(u, args):
+    controller = args.controller
+    antipassback = args.antipassback
+
+    if antipassback == 'disabled':
+        u.set_antipassback(controller, 0)
+    elif antipassback == '(1:2);(3:4)':
+        u.set_antipassback(controller, 1)
+    elif antipassback == '(1,3):(2,4)':
+        u.set_antipassback(controller, 2)
+    elif antipassback == '1:(2,3)':
+        u.set_antipassback(controller, 3)
+    elif antipassback == '1:(2,3,4)':
+        u.set_antipassback(controller, 4)
+    else:
+        raise Exception(f"invalid anti-passback mode {antipassback}")
+
+    display('set-antipassback', [
+        ('ID', controller),
+        ('anti-passback', antipassback),
     ])