+++ tests/ccl/packages.lisp
@@ -36,6 +36,8 @@
            set-interlock
            activate-keypads
            set-door-passcodes
+           get-antipassback
+           set-antipassback
            restore-default-parameters
            listen-events
            internationalisation