+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+1. Implemented `get/set-antipassback` API function to manage the anti-passback mode for a controller.
+
 
 ## [0.8.10](https://github.com/uhppoted/uhppoted-dll/releases/tag/v0.8.10) - 2025-01-30